+++ inc/opendb.php
@@ -21,7 +21,8 @@ if(strtoupper(getenv('PLATFORM')) === "CLOUD"){
 	$allowed_ext = getenv('FILE_EXT') ?: "pdf, doc, docx, xls, csv, xlsx, png, jpg, jpeg, gif";
 	$max_filesize = getenv('MAX_FILE_SIZE') ?: "4194304";
 	$bkparams =  getenv('DB_BACKUP_PARAMETERS') ?: '--column-statistics=1';
-	
+	$BKPOD = getenv('BKPOD') ?: 'localhost';
+
 	$conn = dbConnect($dbhost, $dbuser, $dbpass, $dbname);
 
 }elseif(file_exists(__ROOT__.'/inc/config.php') == TRUE) {